Added persistence map.
This map is not really functional anymore due to the persistence
functions and the exception handling.  But at least it works :).

An set_field call will automatically trigger the write process of the
loader function and will automatically store the new subtree in it.  The
usage of the map or the tree is still the same, just the backend is
differnt now.

Here is an example of a persistence tree with this properties:
* Map size:  1,000,000 x 1,000,000
* Field type: String
* Default field in map: ""
* Default field outside of map "X"
* Save results in 512x512 fields
* Filename is mymap_<x>_<y>_<depth>.map

let map = create_persistence_map (1000000 1000000) "" "X" 10 "mymap_"